# add OSPlatform.macOS, switch to OrdinalIgnoreCase for OSPlatform comparisons