--- a/src/libraries/System.Runtime.InteropServices.RuntimeInformation/src/System/Runtime/InteropServices/RuntimeInformation/RuntimeInformation.Unix.cs
+++ b/src/libraries/System.Runtime.InteropServices.RuntimeInformation/src/System/Runtime/InteropServices/RuntimeInformation/RuntimeInformation.Unix.cs
@@ -21,7 +21,7 @@ namespace System.Runtime.InteropServices
             {
                 return true;
             }
-            else if (name.Length == 3 && name[0] == 'O' && name[1] == 'S' && name[2] == 'X')
+            else if (name == "OSX")
             {
                 return osPlatform.Equals("MACOS", StringComparison.OrdinalIgnoreCase);
             }