--- a/src/libraries/System.Runtime.InteropServices.RuntimeInformation/tests/PlatformVersionTests.cs
+++ b/src/libraries/System.Runtime.InteropServices.RuntimeInformation/tests/PlatformVersionTests.cs
@@ -14,6 +14,11 @@ namespace System.Runtime.InteropServices.RuntimeInformationTests
             yield return new object[] { OSPlatform.Linux };
             yield return new object[] { OSPlatform.OSX };
             yield return new object[] { OSPlatform.Browser };
+            yield return new object[] { OSPlatform.macOS };
+            yield return new object[] { OSPlatform.iOS };
+            yield return new object[] { OSPlatform.tvOS };
+            yield return new object[] { OSPlatform.watchOS };
+            yield return new object[] { OSPlatform.Android };
         }
 
         [Fact]
@@ -49,6 +54,8 @@ namespace System.Runtime.InteropServices.RuntimeInformationTests
             Version current = Environment.OSVersion.Version;
 
             Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater($"{osPlatform}{current}"));
+            Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater($"{osPlatform.ToString().ToLower()}{current}"));
+            Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater($"{osPlatform.ToString().ToUpper()}{current}"));
 
             Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater(osPlatform, current.Major));
 
@@ -78,6 +85,8 @@ namespace System.Runtime.InteropServices.RuntimeInformationTests
 
             Version newer = new Version(currentVersion.Major + 1, 0);
             Assert.False(RuntimeInformation.IsOSPlatformOrLater($"{osPlatform}{newer}"));
+            Assert.False(RuntimeInformation.IsOSPlatformOrLater($"{osPlatform.ToString().ToLower()}{newer}"));
+            Assert.False(RuntimeInformation.IsOSPlatformOrLater($"{osPlatform.ToString().ToUpper()}{newer}"));
             Assert.False(RuntimeInformation.IsOSPlatformOrLater(osPlatform, newer.Major));
 
             newer = new Version(currentVersion.Major, currentVersion.Minor + 1);
@@ -104,6 +113,8 @@ namespace System.Runtime.InteropServices.RuntimeInformationTests
 
             Version older = new Version(current.Major - 1, 0);
             Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater($"{osPlatform}{older}"));
+            Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater($"{osPlatform.ToString().ToLower()}{older}"));
+            Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater($"{osPlatform.ToString().ToUpper()}{older}"));
             Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformOrLater(osPlatform, older.Major));
 
             if (current.Minor > 0)
@@ -160,6 +171,8 @@ namespace System.Runtime.InteropServices.RuntimeInformationTests
             Version current = Environment.OSVersion.Version;
 
             Assert.False(RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform}{current}"));
+            Assert.False(RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform.ToString().ToLower()}{current}"));
+            Assert.False(RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform.ToString().ToUpper()}{current}"));
 
             Assert.False(RuntimeInformation.IsOSPlatformEarlierThan(osPlatform, current.Major));
 
@@ -190,6 +203,8 @@ namespace System.Runtime.InteropServices.RuntimeInformationTests
 
             Version newer = new Version(current.Major + 1, 0);
             Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform}{newer}"));
+            Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform.ToString().ToLower()}{newer}"));
+            Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform.ToString().ToUpper()}{newer}"));
             Assert.Equal(isCurrentPlatfom, RuntimeInformation.IsOSPlatformEarlierThan(osPlatform, newer.Major));
 
             newer = new Version(current.Major, current.Minor + 1);
@@ -215,6 +230,8 @@ namespace System.Runtime.InteropServices.RuntimeInformationTests
 
             Version older = new Version(current.Major - 1, 0);
             Assert.False(RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform}{older}"));
+            Assert.False(RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform.ToString().ToLower()}{older}"));
+            Assert.False(RuntimeInformation.IsOSPlatformEarlierThan($"{osPlatform.ToString().ToUpper()}{older}"));
             Assert.False(RuntimeInformation.IsOSPlatformEarlierThan(osPlatform, older.Major));
 
             if (current.Minor > 0)