I tried to introduce as small overhead as possible.

On Windows, there is basically no difference. On OSX there is no difference for "OSX" (old check), but the "macOS" check takes twice the time because we need two comparisons.

I've added the test cases that were missing in #39005 (ignore case comparison was not enabled then)

## Windows

```cs
private readonly OSPlatform _notCurrent = OSPlatform.Create("NOPE");
private readonly OSPlatform _current = OSPlatform.Windows;

[Benchmark]
public bool IsOSPlatform_True() => RuntimeInformation.IsOSPlatform(_current);

[Benchmark]
public bool IsOSPlatform_False() => RuntimeInformation.IsOSPlatform(_notCurrent);
```

```ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.18363.900 (1909/November2018Update/19H2)
Intel Xeon CPU E5-1650 v4 3.60GHz, 1 CPU, 12 logical and 6 physical cores
```

Before:

|             Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------:|------:|------:|----------:|
|  IsOSPlatform_True | 3.160 ns | 0.0116 ns | 0.0103 ns | 3.158 ns | 3.145 ns | 3.183 ns |     - |     - |     - |         - |
| IsOSPlatform_False | 4.592 ns | 0.0139 ns | 0.0130 ns | 4.590 ns | 4.571 ns | 4.613 ns |     - |     - |     - |         - |

After:

|             Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------:|------:|------:|----------:|
|  IsOSPlatform_True | 3.177 ns | 0.0210 ns | 0.0197 ns | 3.171 ns | 3.152 ns | 3.213 ns |     - |     - |     - |         - |
| IsOSPlatform_False | 4.739 ns | 0.0126 ns | 0.0112 ns | 4.741 ns | 4.722 ns | 4.761 ns |     - |     - |     - |         - |

### OSX

```ini
BenchmarkDotNet=v0.12.1, OS=macOS Mojave 10.14.5 (18F132) [Darwin 18.6.0]
Intel Core i7-5557U CPU 3.10GHz (Broadwell), 1 CPU, 4 logical and 2 physical cores
```

```cs
private readonly OSPlatform _notCurrent = OSPlatform.Create("NOPE");
private readonly OSPlatform _current = OSPlatform.OSX;
private readonly OSPlatform _macOS = OSPlatform.macOS;

[Benchmark]
public bool IsOSPlatform_True() => RuntimeInformation.IsOSPlatform(_current);

[Benchmark]
public bool IsOSPlatform_macOS() => RuntimeInformation.IsOSPlatform(_macOS );

[Benchmark]
public bool IsOSPlatform_False() => RuntimeInformation.IsOSPlatform(_notCurrent);
```

Before:

|             Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------:|------:|------:|----------:|
|  IsOSPlatform_True | 2.901 ns | 0.0071 ns | 0.0063 ns | 2.901 ns | 2.892 ns | 2.915 ns |     - |     - |     - |         - |
| IsOSPlatform_macOS | 4.091 ns | 0.0092 ns | 0.0077 ns | 4.089 ns | 4.081 ns | 4.106 ns |     - |     - |     - |         - |
| IsOSPlatform_False | 4.415 ns | 0.0209 ns | 0.0163 ns | 4.411 ns | 4.402 ns | 4.463 ns |     - |     - |     - |         - |

After:

|             Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------:|------:|------:|----------:|
|  IsOSPlatform_True | 2.887 ns | 0.0057 ns | 0.0048 ns | 2.886 ns | 2.880 ns | 2.897 ns |     - |     - |     - |         - |
| IsOSPlatform_macOS | 7.972 ns | 0.0146 ns | 0.0130 ns | 7.968 ns | 7.956 ns | 7.996 ns |     - |     - |     - |         - |
| IsOSPlatform_False | 9.689 ns | 0.0378 ns | 0.0335 ns | 9.682 ns | 9.627 ns | 9.751 ns |     - |     - |     - |         - |